+++ .env.example
@@ -6,6 +6,7 @@ SMF_BOT_MATRIX_BOT_USER_ID="@NotValidExampleAccount:matrix.org"
 SMF_BOT_NETWORK_DECIMALS=12
 SMF_BOT_NETWORK_UNIT="UNIT"
 SMF_BOT_FAUCET_IGNORE_LIST=""
+SMF_BOT_MATRIX_SERVER=https://matrix.parity.io
 
 # BACKEND
 SMF_BACKEND_FAUCET_ACCOUNT_MNEMONIC="this is a fake mnemonic"
