+++ src/server/checkHealth.ts
@@ -1,14 +1,14 @@
 import { Request, Response } from 'express';
 
 import { logger } from '../utils';
-import apiInstance from './rpc';
+import rpcApiInstance from './lib/rpcApiInstance';
 
 export const checkHealth = async (
   _req: Request,
   res: Response
 ): Promise<void> => {
   try {
-    await apiInstance.isReady;
+    await rpcApiInstance.isReady;
     res.status(200).send({ msg: 'Faucet backend is healthy.' });
   } catch (e) {
     logger.error(`⭕ Api error: ${(e as Error).message}`);
