+++ src/server/constants.ts
@@ -1,0 +1,13 @@
+import { MetricsDefinition } from '../types';
+
+export const metricsDefinition: MetricsDefinition = {
+  data: {
+    errors_rpc_timeout: 0,
+    errors_total: 0,
+    success_requests: 0,
+    total_requests: 0,
+  },
+  meta: {
+    prefix: 'faucet',
+  },
+};
