+++ src/server/lib/ErrorCounter.ts
